requirements.txt
altgraph==0.16.1
appdirs==1.4.3
attrs==19.1.0
Automat==20.2.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==4.0.2
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==1.11.11
django-debug-toolbar==2.1
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
fake-useragent==0.1.11
Flask==1.1.1
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.3
numpy==1.17.1
pandas==0.25.1
pandocfilters==1.4.2
paramiko==2.7.1
parse==1.12.1
parsel==1.5.2
parso==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.6
pycryptodomex==3.9.6
PyDispatcher==2.0.5
pyee==6.0.0
PyExecJS==1.5.1
Pygments==2.4.2
PyHamcrest==2.0.0
PyInstaller==3.5
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
queuelib==1.5.0
redis==3.4.1
requests==2.22.0
requests-html==0.10.0
scipy==1.3.1
Scrapy==1.8.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.5
splinter==0.13.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.40.0
traitlets==4.3.2
Twisted==19.10.0
urllib3==1.25.3
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.5.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.7.1
